requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.0
django-crontab==0.7.1
django-filter==25.2
djangorestframework==3.17.1
gunicorn==25.2.0
idna==3.11
Markdown==3.10.2
packaging==26.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
